Move copyright line to tahoni lib
