+++ projects/tahoni-lib/src/lib/components/copyright.component.html
@@ -1,0 +1,13 @@
+<div class="row copyright">
+  <div class="col-md-2">
+    <fa-icon class="copyright-icon" *ngFor="let tahoniIcon of copyrightService.getCopyrightIcons()" [icon]="tahoniIcon"></fa-icon>
+  </div>
+  <div class="col-md-10 align-items-center justify-content-end" style="display: flex">
+    <div style="text-align: right">
+      <h2>
+        <fa-icon class="copyright-text" [icon]="copyrightService.getCopyrightSymbol()"></fa-icon>
+        <span class="copyright-text">&nbsp;{{copyrightService.getCopyrightText()}}</span>
+      </h2>
+    </div>
+  </div>
+</div>
