+++ projects/tahoni-lib/src/lib/components/copyright.component.scss
@@ -1,5 +1,7 @@
 /* Global styling */
-@import "/dist/tahoni-lib/src/assets/css/colors";
+@import "../../assets/css/colors";
+
+/* Component-specific styling */
 
 /* Colours */
 .copyright-icon, .copyright-text {
