--- projects/tahoni-lib/src/lib/components/copyright/copyright/copyright.component.html
@@ -1,1 +1,0 @@
-<p>copyright works!</p>
