--- projects/tahoni-lib/src/lib/components/copyright/copyright/copyright.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'lib-copyright',
-  templateUrl: './copyright.component.html',
-  styleUrls: ['./copyright.component.css']
-})
-export class CopyrightComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
