+++ projects/tahoni-lib/src/lib/services/copyright.service.ts
@@ -1,13 +1,18 @@
 import {Injectable} from '@angular/core';
-import {faCat, faMouse} from '@fortawesome/free-solid-svg-icons';
+import {faCat, faCopyright, faMouse} from '@fortawesome/free-solid-svg-icons';
 
 @Injectable({
   providedIn: 'root'
 })
 export class CopyrightService {
 
+  private copyrightSymbol =  faCopyright;
   private copyrightLogo = [faCat, faMouse];
   private copyrightText = '2021 tahoni';
+
+  getCopyrightSymbol() {
+    return this.copyrightSymbol;
+  }
 
   getCopyrightIcons() {
     return this.copyrightLogo;
