+++ projects/tahoni-lib/src/lib/tahoni-lib.module.ts
@@ -1,5 +1,10 @@
 import {NgModule} from '@angular/core';
 
+import {CommonModule} from '@angular/common';
+import {FontAwesomeModule} from '@fortawesome/angular-fontawesome';
+
+import {CopyrightComponent} from './components/copyright.component';
+
 import {FocusOnErrorDirective} from './directives/focus-on-error.directive';
 import {TouchOnErrorDirective} from './directives/touch-on-error.directive';
 import {MessageOnErrorDirective} from './directives/message-on-error.directive';
@@ -7,7 +12,6 @@ import {SummaryOnErrorDirective} from './directives/summary-on-error.directive';
 
 import {AlertService} from './services/alert.service';
 import {CopyrightService} from './services/copyright.service';
-import {CopyrightComponent} from './components/copyright/copyright/copyright.component';
 
 @NgModule({
   declarations: [
@@ -17,12 +21,16 @@ import {CopyrightComponent} from './components/copyright/copyright/copyright.com
     SummaryOnErrorDirective,
     CopyrightComponent,
   ],
-  imports: [],
+  imports: [
+    CommonModule,
+    FontAwesomeModule
+  ],
   exports: [
     FocusOnErrorDirective,
     TouchOnErrorDirective,
     MessageOnErrorDirective,
     SummaryOnErrorDirective,
+    CopyrightComponent,
   ],
   providers: [
     AlertService,
