+++ projects/tahoni-lib/src/projects.ts
@@ -6,6 +6,7 @@ export * from './lib/directives/focus-on-error.directive';
 export * from './lib/directives/touch-on-error.directive';
 export * from './lib/directives/message-on-error.directive';
 export * from './lib/directives/summary-on-error.directive';
+export * from './lib/components/copyright.component'
 export * from './lib/services/alert.service';
 export * from './lib/services/copyright.service';
